@@ -1,8 +1,8 @@
 module github.com/mheers/kubeyaml
 
-go 1.19
+go 1.20
 
 require (
 	github.com/pkg/errors v0.9.1
-	gopkg.in/yaml.v2 v2.3.0
+	gopkg.in/yaml.v2 v2.4.0
 )
