@@ -3,7 +3,7 @@ package main
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 )
@@ -36,6 +36,7 @@ func main() {
 		"OneTwentyThree": "1.23",
 		"OneTwentyFour":  "1.24",
 		"OneTwentyFive":  "1.25",
+		"OneTwentySix":   "1.26",
 	}
 	urlFmt := "https://raw.githubusercontent.com/kubernetes/kubernetes/release-%s/api/openapi-spec/swagger.json"
 
@@ -46,7 +47,7 @@ func main() {
 			continue
 		}
 
-		schema, err := ioutil.ReadAll(r.Body)
+		schema, err := io.ReadAll(r.Body)
 		if err != nil {
 			fmt.Println("headers", r.Header)
 			fmt.Println("status code:", r.StatusCode)
@@ -60,14 +61,14 @@ func main() {
 			continue
 		}
 
-		if err := ioutil.WriteFile(fmt.Sprintf("%s/internal/kubernetes/data/swagger-%s.json", outDir, release), schema, os.FileMode(uint32(0644))); err != nil {
+		if err := os.WriteFile(fmt.Sprintf("%s/internal/kubernetes/data/swagger-%s.json", outDir, release), schema, os.FileMode(uint32(0644))); err != nil {
 			fmt.Printf("error writing file release-%s: %v", release, err)
 			continue
 		}
 		schema = bytes.Replace(schema, []byte("`"), []byte("` + \"`\" + `"), -1)
 		fileData := fmt.Sprintf(templateFormat, funcName, schema)
 
-		if err := ioutil.WriteFile(fmt.Sprintf("%s/internal/kubernetes/data/swagger_%s.go", outDir, funcName), []byte(fileData), os.FileMode(uint32(0644))); err != nil {
+		if err := os.WriteFile(fmt.Sprintf("%s/internal/kubernetes/data/swagger_%s.go", outDir, funcName), []byte(fileData), os.FileMode(uint32(0644))); err != nil {
 			fmt.Printf("error writing go file for release-%s: %v", release, err)
 			continue
 		}
